Fix file path handeling
